Initial Release of Python Install Generator
This commit contains the initial version of the Python script
"generate_installer_sql.py", which takes the .SQL files under the
Code/ subfolders and concatenates them into a single release file.

This Python script is controlled via a config file at
Installation/config, and config template files are available in the
Installation/ folder.

The CoreXR files have been modified in some cases to cooperate with
the new installation process, and many small bugs were fixed.

The file tmp_CleanUpChiRho.sql has been added, but this is just a
convenience script that will be reworked in the future.